# Faster ReadString by avoiding memory allocations in []byte to string conversion