--- a/b2s_new.go
+++ b/b2s_new.go
@@ -1,0 +1,12 @@
+//go:build go1.20
+// +build go1.20
+
+package jsoniter
+
+import "unsafe"
+
+// b2s converts byte slice to a string without memory allocation.
+// See https://groups.google.com/forum/#!msg/Golang-Nuts/ENgbUzYvCuU/90yGx7GUAgAJ .
+func b2s(b []byte) string {
+	return unsafe.String(unsafe.SliceData(b), len(b))
+}