--- a/b2s_old.go
+++ b/b2s_old.go
@@ -1,0 +1,15 @@
+//go:build !go1.20
+// +build !go1.20
+
+package jsoniter
+
+import "unsafe"
+
+// b2s converts byte slice to a string without memory allocation.
+// See https://groups.google.com/forum/#!msg/Golang-Nuts/ENgbUzYvCuU/90yGx7GUAgAJ .
+//
+// Note it may break if string and/or slice header will change
+// in the future go versions.
+func b2s(b []byte) string {
+	return *(*string)(unsafe.Pointer(&b))
+}