--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/json-iterator/go
 
-go 1.12
+go 1.20
 
 require (
 	github.com/davecgh/go-spew v1.1.1
@@ -8,4 +8,9 @@ require (
 	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421
 	github.com/modern-go/reflect2 v1.0.2
 	github.com/stretchr/testify v1.8.0
+)
+
+require (
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )